Removed 'Stream not initialized' check
Checking _reading is not correct, because
it will be set to true only if the stream
is in the flowing mode.  The stream consumer
may choose to use the stream in the paused mode,
and call the read() function to iterativelly
get each next result, in which case the check
will fail.