Fix error handling
When an error is emitted by the pool, before pg-promise received the
client, the error handling failed with:

> TypeError: Cannot read property 'options' of undefined

Try to log pool errors best effort.

---

When an in-use client receives an error rely on the caller to end the
connection instead of doing it in the error handler to prevent releasing
the same connection twice.